+++ CHANGELOG.md
@@ -15,6 +15,7 @@ For a full diff see [`5.16.0...main`][5.16.0...main].
 - Extracted `Config\Rules` as a value object ([#874]), by [@localheinz]
 - Extracted `Config\Fixers` as a value object ([#883]), by [@localheinz]
 - Added `Config\RuleSet::withRules()` to allow overriding `Config\Rules` ([#891]), by [@localheinz]
+- Added `Config\RuleSet::withCustomFixers()` to allow registering additional `Config\Fixers` ([#893]), by [@localheinz]
 
 ### Changed
 
@@ -1206,6 +1207,7 @@ For a full diff see [`d899e77...1.0.0`][d899e77...1.0.0].
 [#888]: https://github.com/ergebnis/php-cs-fixer-config/pull/888
 [#891]: https://github.com/ergebnis/php-cs-fixer-config/pull/891
 [#892]: https://github.com/ergebnis/php-cs-fixer-config/pull/892
+[#893]: https://github.com/ergebnis/php-cs-fixer-config/pull/893
 
 [@dependabot]: https://github.com/apps/dependabot
 [@linuxjuggler]: https://github.com/linuxjuggler